chore: move password and srp list items
