+++ ui/pages/settings-v2/security-and-password-tab/change-password/change-password-warning.tsx
