+++ ui/pages/settings-v2/security-and-password-tab/password-sub-page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import ChangePassword from '../../settings/security-tab/change-password/change-password';
 import { SECURITY_AND_PASSWORD_ROUTE } from '../../../helpers/constants/routes';
+import ChangePassword from './change-password';
 
 const PasswordSubPage = () => {
   return <ChangePassword redirectRoute={SECURITY_AND_PASSWORD_ROUTE} />;
