+++ ui/pages/settings-v2/security-and-password-tab/reveal-srp-list/index.ts
